+++ Assets/scripts/Invisible.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b94895586a061bf41909fbbe8526acd1
+guid: c1f69f8f3f6d46a41aaf762bc18cb15b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
