+++ Assets/scripts/InvisibleScriptableObject.cs.meta
