+++ Assets/scripts/TargetColorChange.cs
@@ -2,11 +2,11 @@ using System.Collections;
 using UnityEngine;
 
 /* 
- * Written this semester,  starting at line 13.
+ * Written this semester, starting at line 13.
  * This script handles the current targets color and notifies the target script manager when it is fired at.
  * It is also prepared to be reset when needed.
  */
-//This one line was generated by unity
+//This one line was generated by unity.
 public class TargetColorChange : MonoBehaviour
 {
     //Sets the targets color to a new color from the target script manager.
@@ -14,16 +14,16 @@ public class TargetColorChange : MonoBehaviour
     {
         //Get the mesh renderer and the material.
         MeshRenderer mesh = GetComponent<MeshRenderer>();
-        //Get the next avalable color from the target scrip manager.
+        //Get the next available color from the target scrip manager.
         Material material = TargetScriptManager.instance.availableTargetMaterial.Pop();
-        //Set the target's color to the next color
+        //Set the target's color to the next color.
         mesh.sharedMaterial = material;
     }
     public void ResetTarget()
     {
-        //Seet a new color
+        //Set a new color.
         SetColor();
-        //To flip the target back the trigger has to be bent back. So tell the triggers script to change position.
+        //To flip the target back the trigger must be bent back. So, tell the triggers script to change position.
         StopCoroutine(GetComponent<trigger>().FlipTarget());
         //Reset the target by getting the trigger script and rotating the target back to position.
         StartCoroutine(GetComponent<trigger>().FlipTargetBack());
@@ -44,7 +44,7 @@ public class TargetColorChange : MonoBehaviour
     //Called by the gun when fired at.
     public void TargitHit(float _)
     {
-        //Start the ResisterHit Coroutine
+        //Start the RegisterHit Coroutine.
         StartCoroutine(RegisterHit());
     }
