+++ Assets/scripts/TargetScriptManager.cs
@@ -2,29 +2,29 @@ using System.Collections.Generic;
 using UnityEngine;
 
 /* 
- * Written this semester,  starting at line 12.
- * This script handles the target's colors and keeps track of which order they have to be fired at to get the reward. It also keeps track of the reward.
+ * Written this semester, starting at line 12.
+ * This script handles the target's colors and keeps track of which order they must be fired at to get the reward. It also keeps track of the reward.
  */
 //This one line was generated by unity
 public class TargetScriptManager : MonoBehaviour
 {
-    //Setting up a singleton
+    //Setting up a singleton.
     public static TargetScriptManager instance;
     //Events for the targets to subscribe to.
     public event System.Action onReset = () => { };
     public event System.Action newColor = () => { };
 
-    //The colors of the targets, set in the inspector
+    //The colors of the targets, set in the inspector.
     public List<Material> materials = new();
-    //A stack which represents the order in which the targets have to be fired at.
+    //A stack which represents the order in which the targets must be fired at.
     public Stack<Material> targetColorOrder = new();
-    //This stack represents the colors. The targets pop a color off of this stack and use it for it's own color.
+    //This stack represents the colors. The targets pop a color off this stack and use it for its own color.
     public Stack<Material> availableTargetMaterial = new();
 
-    //The reward recived when the player fire the targes in the correct order.
+    //The reward received when the player fires the targets in the correct order.
     public GameObject gameObjectToEnable;
 
-    //Suffles the colors of the targets and resets the targetColorOrder stack to the way it was when the gam started.
+    //Shuffles the colors of the targets and resets the targetColorOrder stack to the way it was when the game started.
     public void ResetTarget()
     {
         availableTargetMaterial.Clear();
@@ -51,14 +51,14 @@ public class TargetScriptManager : MonoBehaviour
 
     public void Awake()
     {
-        //Setting up the singleton
+        //Setting up the singleton.
         instance = this;
     }
 
-    // Start is called before the first frame update
+    // Start is called before the first frame update.
     void Start()
     {
-        //Suffle the materials list. From now on the order of this list determines which order the targets must be fired at.
+        //Shuffle the materials list. From now on the order of this list determines which order the targets must be fired at.
         materials.Rearrange();
         //Fill the other stacks and set up for the game.
         ResetTarget();
@@ -73,7 +73,7 @@ public class TargetScriptManager : MonoBehaviour
         newColor();
     }
 
-    //Called by a target when the taget has been hit.
+    //Called by a target when the target has been hit.
     public void FiredAtTarget(Material self)
     {
         //If the targetColorOrder is empty than the player is erroneously firing at hit targets.
@@ -81,18 +81,18 @@ public class TargetScriptManager : MonoBehaviour
         {
             //Pick the next color off the stack. This is the next color that is supposed to be hit.
             Material target = targetColorOrder.Pop();
-            //If the materail hit is not the next on that is supposed to be hit than reset the targets.
+            //If the material hit is not the next on that is supposed to be hit than reset the targets.
             if (self != target)
             {
                 ResetTarget();
                 onReset();
             }
-            //If the correct materail has hit and the stack is now empty than the player has fired at all of the targets in the correct order.
+            //If the correct material has hit and the stack is now empty than the player has fired at all the targets in the correct order.
             else if (targetColorOrder.Count == 0)
             {
                 //Activate the reward object.
                 gameObjectToEnable.SetActive(true);
-                //Clear the onReset event. This puzzle is finsihed.
+                //Clear the onReset event. This puzzle is finished.
                 onReset = () => { };
             }
         }
